feat: moved purge services to gateway
fixes: AM-4567
