+++ gravitee-am-dataplane/gravitee-am-dataplane-api/src/main/java/io/gravitee/am/dataplane/api/repository/LoginAttemptRepository.java
@@ -18,6 +18,7 @@ package io.gravitee.am.dataplane.api.repository;
 import io.gravitee.am.dataplane.api.search.LoginAttemptCriteria;
 import io.gravitee.am.model.LoginAttempt;
 import io.gravitee.am.repository.common.CrudRepository;
+import io.gravitee.am.repository.common.ExpiredDataSweeper;
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Maybe;
 
@@ -25,13 +26,10 @@ import io.reactivex.rxjava3.core.Maybe;
  * @author Titouan COMPIEGNE (titouan.compiegne at graviteesource.com)
  * @author GraviteeSource Team
  */
-public interface LoginAttemptRepository extends CrudRepository<LoginAttempt, String> {
+public interface LoginAttemptRepository extends CrudRepository<LoginAttempt, String>, ExpiredDataSweeper {
 
     Maybe<LoginAttempt> findByCriteria(LoginAttemptCriteria criteria);
 
     Completable delete(LoginAttemptCriteria criteria);
 
-    default Completable purgeExpiredData() {
-        return Completable.complete();
-    }
 }
