+++ gravitee-am-gateway/gravitee-am-gateway-services/gravitee-am-gateway-services-purge/src/main/assembly/plugin-assembly.xml
