+++ gravitee-am-gateway/gravitee-am-gateway-services/pom.xml
@@ -32,5 +32,6 @@
     <packaging>pom</packaging>
     <modules>
         <module>gravitee-am-gateway-services-sync</module>
+        <module>gravitee-am-gateway-services-purge</module>
     </modules>
 </project>
