+++ gravitee-am-gateway/gravitee-am-gateway-standalone/gravitee-am-gateway-standalone-distribution/pom.xml
@@ -195,6 +195,14 @@
         </dependency>
 
         <dependency>
+            <groupId>io.gravitee.am.gateway</groupId>
+            <artifactId>gravitee-am-gateway-services-purge</artifactId>
+            <version>${project.version}</version>
+            <type>zip</type>
+            <scope>runtime</scope>
+        </dependency>
+
+        <dependency>
             <groupId>io.gravitee.am.reporter</groupId>
             <artifactId>gravitee-am-reporter-mongodb</artifactId>
             <version>${project.version}</version>
@@ -621,6 +629,13 @@
                                 <artifactItem>
                                     <groupId>io.gravitee.am.gateway</groupId>
                                     <artifactId>gravitee-am-gateway-services-sync</artifactId>
+                                    <version>${project.version}</version>
+                                    <type>zip</type>
+                                </artifactItem>
+
+                                <artifactItem>
+                                    <groupId>io.gravitee.am.gateway</groupId>
+                                    <artifactId>gravitee-am-gateway-services-purge</artifactId>
                                     <version>${project.version}</version>
                                     <type>zip</type>
                                 </artifactItem>
