+++ gravitee-am-management-api/gravitee-am-management-api-services/pom.xml
@@ -33,6 +33,5 @@
 
     <modules>
         <module>gravitee-am-management-api-services-sync</module>
-        <module>gravitee-am-management-api-services-purge</module>
     </modules>
 </project>
