+++ gravitee-am-management-api/gravitee-am-management-api-standalone/gravitee-am-management-api-standalone-distribution/pom.xml
@@ -221,14 +221,6 @@
         </dependency>
 
         <dependency>
-            <groupId>io.gravitee.am.management</groupId>
-            <artifactId>gravitee-am-management-api-services-purge</artifactId>
-            <version>${project.version}</version>
-            <type>zip</type>
-            <scope>runtime</scope>
-        </dependency>
-
-        <dependency>
             <groupId>io.gravitee.am.factor</groupId>
             <artifactId>gravitee-am-factor-otp</artifactId>
             <version>${project.version}</version>
@@ -638,13 +630,6 @@
                                 <artifactItem>
                                     <groupId>io.gravitee.am.management</groupId>
                                     <artifactId>gravitee-am-management-api-services-sync</artifactId>
-                                    <version>${project.version}</version>
-                                    <type>zip</type>
-                                </artifactItem>
-
-                                <artifactItem>
-                                    <groupId>io.gravitee.am.management</groupId>
-                                    <artifactId>gravitee-am-management-api-services-purge</artifactId>
                                     <version>${project.version}</version>
                                     <type>zip</type>
                                 </artifactItem>
