+++ gravitee-am-plugins-handlers/gravitee-am-plugins-handlers-dataplane/src/main/java/io/gravitee/am/plugins/dataplane/core/DataPlaneRegistry.java
@@ -47,6 +47,8 @@ public interface DataPlaneRegistry {
 
     DataPlaneDescription getDescription(Domain domain);
 
+    DataPlaneProvider getProviderById(String id);
+
     CredentialRepository getCredentialRepository(Domain domain);
 
     DeviceRepository getDeviceRepository(Domain domain);
