+++ gravitee-am-repository/gravitee-am-repository-api/src/main/java/io/gravitee/am/repository/management/api/PermissionTicketRepository.java
@@ -17,15 +17,13 @@ package io.gravitee.am.repository.management.api;
 
 import io.gravitee.am.model.uma.PermissionTicket;
 import io.gravitee.am.repository.common.CrudRepository;
+import io.gravitee.am.repository.common.ExpiredDataSweeper;
 import io.reactivex.rxjava3.core.Completable;
 
 /**
  * @author Alexandre FARIA (contact at alexandrefaria.net)
  * @author GraviteeSource Team
  */
-public interface PermissionTicketRepository  extends CrudRepository<PermissionTicket, String> {
+public interface PermissionTicketRepository extends CrudRepository<PermissionTicket, String>, ExpiredDataSweeper {
 
-    default Completable purgeExpiredData() {
-        return Completable.complete();
-    }
 }
